introduced LongLiteralExp consistently in OCL, UML, Ecore; one test fails

+++ plugins/org.eclipse.ocl.ecore/src/org/eclipse/ocl/ecore/impl/IntegerLiteralExpImpl.java
@@ -55,7 +55,7 @@ public class IntegerLiteralExpImpl
 	 * @generated
 	 * @ordered
 	 */
-	protected static final Long INTEGER_SYMBOL_EDEFAULT = null;
+	protected static final Integer INTEGER_SYMBOL_EDEFAULT = null;
 
 	/**
 	 * The cached value of the '{@link #getIntegerSymbol() <em>Integer Symbol</em>}' attribute.
@@ -65,7 +65,7 @@ public class IntegerLiteralExpImpl
 	 * @generated
 	 * @ordered
 	 */
-	protected Long integerSymbol = INTEGER_SYMBOL_EDEFAULT;
+	protected Integer integerSymbol = INTEGER_SYMBOL_EDEFAULT;
 
 	/**
 	 * <!-- begin-user-doc -->
@@ -90,9 +90,8 @@ public class IntegerLiteralExpImpl
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @generated
-	 * @since 3.2
 	 */
-	public Long getIntegerSymbol() {
+	public Integer getIntegerSymbol() {
 		return integerSymbol;
 	}
 
@@ -100,10 +99,9 @@ public class IntegerLiteralExpImpl
 	 * <!-- begin-user-doc -->
 	 * <!-- end-user-doc -->
 	 * @generated
-	 * @since 3.2
 	 */
-	public void setIntegerSymbol(Long newIntegerSymbol) {
-		Long oldIntegerSymbol = integerSymbol;
+	public void setIntegerSymbol(Integer newIntegerSymbol) {
+		Integer oldIntegerSymbol = integerSymbol;
 		integerSymbol = newIntegerSymbol;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET,
@@ -145,7 +143,7 @@ public class IntegerLiteralExpImpl
 	public void eSet(int featureID, Object newValue) {
 		switch (featureID) {
 			case EcorePackage.INTEGER_LITERAL_EXP__INTEGER_SYMBOL :
-				setIntegerSymbol((Long) newValue);
+				setIntegerSymbol((Integer) newValue);
 				return;
 		}
 		super.eSet(featureID, newValue);
