+++ plugins/org.eclipse.ocl/src/org/eclipse/ocl/expressions/impl/IntegerLiteralExpImpl.java
@@ -18,6 +18,7 @@
 package org.eclipse.ocl.expressions.impl;
 
 import java.util.Map;
+
 import org.eclipse.emf.common.notify.Notification;
 import org.eclipse.emf.common.util.DiagnosticChain;
 import org.eclipse.emf.ecore.EClass;
@@ -52,7 +53,7 @@ public class IntegerLiteralExpImpl<C>
 	 * @generated
 	 * @ordered
 	 */
-	protected static final Long INTEGER_SYMBOL_EDEFAULT = null;
+	protected static final Integer INTEGER_SYMBOL_EDEFAULT = null;
 
 	/**
 	 * The cached value of the '{@link #getIntegerSymbol() <em>Integer Symbol</em>}' attribute.
@@ -62,7 +63,7 @@ public class IntegerLiteralExpImpl<C>
 	 * @generated
 	 * @ordered
 	 */
-	protected Long integerSymbol = INTEGER_SYMBOL_EDEFAULT;
+	protected Integer integerSymbol = INTEGER_SYMBOL_EDEFAULT;
 
 	/**
 	 * <!-- begin-user-doc -->
@@ -88,7 +89,7 @@ public class IntegerLiteralExpImpl<C>
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public Long getIntegerSymbol() {
+	public Integer getIntegerSymbol() {
 		return integerSymbol;
 	}
 
@@ -97,8 +98,8 @@ public class IntegerLiteralExpImpl<C>
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public void setIntegerSymbol(Long newIntegerSymbol) {
-		Long oldIntegerSymbol = integerSymbol;
+	public void setIntegerSymbol(Integer newIntegerSymbol) {
+		Integer oldIntegerSymbol = integerSymbol;
 		integerSymbol = newIntegerSymbol;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET,
@@ -140,7 +141,7 @@ public class IntegerLiteralExpImpl<C>
 	public void eSet(int featureID, Object newValue) {
 		switch (featureID) {
 			case ExpressionsPackage.INTEGER_LITERAL_EXP__INTEGER_SYMBOL :
-				setIntegerSymbol((Long) newValue);
+				setIntegerSymbol((Integer) newValue);
 				return;
 		}
 		super.eSet(featureID, newValue);
